% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_Functions_forPackage.R
\name{boot.meanmatrixstats}
\alias{boot.meanmatrixstats}
\title{Bootstrapping the MeanMatrixStatistic from the evolqg package}
\usage{
boot.meanmatrixstats(d, i, measurement)
}
\arguments{
\item{d}{dataframe}

\item{i}{iteration number that will be provided by the boot package}

\item{measurement}{one of the following statistics: "MeanSquaredCorrelation", "pc1.percent", "ICV","EigenSd","respondability","evolvability","conditional.evolvability", "autonomy", "flexibility", or "constraints"}
}
\value{
boot data frame from which you can make further calculations
}
\description{
This function takes a single dataframe and calculates one of the 10
statistics that the MeanMatrixStatistic function can produce. The user must
provide the data frame and the specific statistics they would like to caculate
}
