% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_Functions_forPackage.R
\name{mean.and.ci}
\alias{mean.and.ci}
\title{Prints the mean and confidence intervals for any of the above functions}
\usage{
\method{mean}{and.ci}(
  enter.dataframe,
  iterations,
  which.function,
  stat = NULL,
  matrix.1.pop = NULL,
  matrix.2.pop = NULL
)
}
\arguments{
\item{enter.dataframe}{dataframe to be bootstrapped. The type of dataframe is dependent upon the function you will use}

\item{iterations}{iteration number that will be provided to the boot package}

\item{which.function}{name of the function you wish to bootstrap: boot.meanmatrixstats, boot.randomskewer, boot.pca.similarity, or boot.matrix.distance (no quotes need here)}

\item{stat}{DEFAULT=NULL, for function with only 1 calculation (PCAsimilarity or MatrixDistance) this parameter can be left blank. For functions that calculate many
different statistics (MeanMatrixStatistic or RandomSkewers), you should enter the name of the stat you'd like to calculate here in quotes (see other function help pages for more info).}

\item{matrix.1.pop}{DEFAULT=NULL, only enter population names (in quotes) for functions that require two populations for calculations.}

\item{matrix.2.pop}{DEFAULT=NULL, only enter population names (in quotes) for functions that require two populations for calculations.}
}
\value{
boot data frame from which you can make further calculations
}
\description{
This function can be used to print the mean and upper & lower confidence intervals for any of the above functions in one go
}
